+++ core/src/main/java/at/campus02/swd/game/observer/PositionObserver.java
@@ -1,5 +1,5 @@
 package at.campus02.swd.game.observer;
 
 public interface PositionObserver {
-    void updatePosition(int x, int y, float rotation);
+    void updatePosition(float x, float y, float rotation);
 }
